use doubles again for x-coordinates. They are needed after all...
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@7411 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/lyxtext.h
@@ -494,10 +494,10 @@ public:
 
 	/** this calculates the specified parameters. needed when setting
 	 * the cursor and when creating a visible row */
-	void prepareToPrint(RowList::iterator row, int & x,
-			    int & fill_separator,
-			    int & fill_hfill,
-			    int & fill_label_hfill,
+	void prepareToPrint(RowList::iterator row, double & x,
+			    double & fill_separator,
+			    double & fill_hfill,
+			    double & fill_label_hfill,
 			    bool bidi = true) const;
 
 private:
